backend/requirements.txt
annotated-types==0.6.0
anthropic==0.25.2
anyio==4.3.0
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cohere==5.3.2
distro==1.9.0
fastavro==1.9.4
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SSE==1.0.0
fsspec==2024.3.1
groq==0.5.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
tokenizers==0.15.2
tqdm==4.66.2
types-requests==2.31.0.20240406
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wsproto==1.2.0
